requirements.txt
Django==1.5.5
PIL==1.1.7
dj-database-url==0.2.2
dj-static==0.0.5
django-crispy-forms==1.3.2
django-toolbelt==0.0.1
gunicorn==18.0
newrelic==2.4.0.4
psycopg2==2.5.1
static==0.4
wadofstuff-django-serializers==1.1.0
wsgiref==0.1.2
